finally done ( ig )

+++ TicketingSystem/TicketingSystem.Application/Services/TicketServices/TicketService.cs
@@ -1,6 +1,7 @@
 using TicketingSystem.Application.Abstractions.IRepositories;
 using TicketingSystem.Application.Abstractions.IServices;
 using TicketingSystem.Domain.Entities.DTOs;
+using TicketingSystem.Domain.Entities.Exceptions;
 using TicketingSystem.Domain.Entities.Models;
 
 namespace TicketingSystem.Application.Services.TicketServices
@@ -8,10 +9,33 @@ namespace TicketingSystem.Application.Services.TicketServices
     public class TicketService : ITicketService
     {
         private readonly ITicketRepository _ticketRepository;
+        private readonly IUserRepository _userRepository;
 
-        public TicketService(ITicketRepository ticketRepository)
+        public TicketService(ITicketRepository ticketRepository, IUserRepository userRepository)
         {
             _ticketRepository = ticketRepository;
+            _userRepository = userRepository;
+        }
+        public async Task<string> PurchaseTicket(string email, string password, int id)
+        {
+            var userByEmail = await _userRepository.GetByAny(x => x.Email == email);
+            var res = await _ticketRepository.GetByAny(x => x.Id == id);
+            if (userByEmail != null)
+            {
+                if (userByEmail.Password == password)
+                {
+                    if (res != null)
+                    {
+                        userByEmail.TicketId = id;
+
+                        await _userRepository.Update(userByEmail);
+                        return "Purchased";
+                    }
+                    return "Such ticket not found";
+                }
+                return "Incorrect password";
+            }
+            return "Incorrect email";
         }
 
         public async Task<string> Create(TicketDTO ticketDTO)
@@ -28,10 +52,24 @@ namespace TicketingSystem.Application.Services.TicketServices
 
         public async Task<bool> Delete(int id)
         {
-            var result = await _ticketRepository.Delete(x => x.Id == id);
+            var ticketById = await _ticketRepository.GetByAny(x => x.Id == id);
+            if (ticketById != null)
+            {
+                var usersWithTicket = await _userRepository.GetByAnyList(user => user.TicketId == id);
 
-            return result;
+                foreach (var user in usersWithTicket)
+                {
+                    user.TicketId = 0;
+                    await _userRepository.Update(user);
+                }
+
+                var result = await _ticketRepository.Delete(x => x.Id == id);
+
+                return result;
+            }
+            throw new TicketNotFoundException("Ticket not found!");
         }
+
 
         public async Task<IEnumerable<Ticket>> GetAll()
         {
@@ -49,25 +87,42 @@ namespace TicketingSystem.Application.Services.TicketServices
         public async Task<Ticket> GetByName(string name)
         {
             var result = await _ticketRepository.GetByAny(d => d.TicketName == name);
-            return result;
+            if (result != null)
+            {
+                return result;
+            }
+            throw new TicketNotFoundException("Ticket not found!");
+        }
+
+        public async Task<Ticket> GetById(int id)
+        {
+            var result = await _ticketRepository.GetByAny(d => d.Id == id);
+            if (result != null)
+            {
+                return result;
+            }
+            throw new TicketNotFoundException("Ticket not found!");
         }
 
         public async Task<string> Update(int Id, TicketDTO ticketDTO)
         {
             var res = await _ticketRepository.GetByAny(x => x.Id == Id);
+            var name = await _ticketRepository.GetByAny(x => x.TicketName == ticketDTO.TicketName);
 
             if (res != null)
             {
-                var Ticket = new Ticket()
+                if (name == null)
                 {
-                    TicketName = ticketDTO.TicketName,
-                    TicketDescription = ticketDTO.TicketDescription
-                };
-                var result = await _ticketRepository.Update(Ticket);
+                    res.TicketName = ticketDTO.TicketName;
+                    res.TicketDescription = ticketDTO.TicketDescription;
 
-                return "Updated";
+                    await _ticketRepository.Update(res);
+
+                    return "Updated";
+                }
+                return "Such ticket already exists";
             }
-            return "Failed";
+            return "Such ticket not found";
 
         }
     }
